+++ server/tests/backends/sql_translator_integration_tests/test_sql_mysql_translator_steps.py
@@ -11,7 +11,7 @@ from docker.types import Ulimit
 from sqlalchemy import create_engine, text
 
 from tests.utils import (
-    _BEERS_TABLE_COLUMNS,
+    BEERS_TABLE_COLUMNS,
     assert_dataframes_equals,
     docker_container,
     get_spec_from_json_fixture,
@@ -87,7 +87,7 @@ def test_sql_translator_pipeline(case_id: str, case_spec_file_path: str, engine:
     query = translate_pipeline(
         sql_dialect=SQLDialect.MYSQL,
         pipeline=pipeline,
-        tables_columns={"beers_tiny": _BEERS_TABLE_COLUMNS},
+        tables_columns={"beers_tiny": BEERS_TABLE_COLUMNS},
         db_schema=None,
     )
     # Execute request generated from Pipeline in Postgres and get the result