+++ server/tests/backends/sql_translator_integration_tests/test_sql_snowflake_translator_steps.py
@@ -1,49 +1,31 @@
 import json
+from datetime import date
 from io import StringIO
 from os import environ
 from typing import Any
+from urllib.parse import quote
 
 import pandas as pd
 import pytest
 from snowflake.sqlalchemy import URL
 from sqlalchemy import create_engine, text
 
-from tests.utils import assert_dataframes_equals, get_spec_from_json_fixture, retrieve_case
+from tests.utils import BEERS_TABLE_COLUMNS, assert_dataframes_equals, get_spec_from_json_fixture, retrieve_case
 from weaverbird.backends.pypika_translator.dialects import SQLDialect
 from weaverbird.backends.pypika_translator.translate import translate_pipeline
 from weaverbird.pipeline import PipelineWithVariables
 from weaverbird.utils.toucan_connectors import nosql_apply_parameters_to_query_with_errors
 
-_ACCOUNT = "toucantocopartner.west-europe.azure"
-_USER = "toucan_test"
-_WAREHOUSE = "toucan_test"
-_DATABASE = "toucan_test"
-_ROLE = "toucan_test"
-_SCHEMA = "toucan_test"
-
-_BEERS_TABLE_COLUMNS = [
-    "price_per_l",
-    "alcohol_degree",
-    "name",
-    "cost",
-    "beer_kind",
-    "volume_ml",
-    "brewing_date",
-    "nullable_name",
-]
-
 
 @pytest.fixture(scope="module")
 def engine():
     url = URL(
-        account=_ACCOUNT,
-        user=_USER,
-        password=environ.get("SNOWFLAKE_PASSWORD"),
-        warehouse=_WAREHOUSE,
-        database=_DATABASE,
-        role=_ROLE,
-        schema=_SCHEMA,
-        authenticator="snowflake",
+        user=environ["SNOWFLAKE_USER"],
+        password=quote(environ["SNOWFLAKE_PASSWORD"]),
+        database=environ["SNOWFLAKE_DATABASE"],
+        account=environ["SNOWFLAKE_ACCOUNT"],
+        schema=environ["SNOWFLAKE_SCHEMA"],
+        warehouse=environ["SNOWFLAKE_WAREHOUSE"],
     )
     engine = create_engine(url)
     connection = engine.connect()
@@ -51,7 +33,14 @@ def engine():
     connection.close()
 
 
-@pytest.skip("Should be skipped, waiting the payment of creds on november...", allow_module_level=True)
+def _sanitize_datetimes(df: pd.DataFrame) -> pd.DataFrame:
+    """Converts date columns to datetime objects"""
+    for col in df.columns:
+        if all(elem is None or isinstance(elem, date) for elem in df[col]):
+            df[col] = pd.to_datetime(df[col])
+    return df
+
+
 @pytest.mark.parametrize("case_id, case_spec_file", retrieve_case("sql_translator", "snowflake_pypika"))
 def test_snowflake_translator_pipeline(engine: Any, case_id: str, case_spec_file: str, available_variables: dict):
     pipeline_spec = get_spec_from_json_fixture(case_id, case_spec_file)
@@ -64,9 +53,14 @@ def test_snowflake_translator_pipeline(engine: Any, case_id: str, case_spec_file
     query = translate_pipeline(
         sql_dialect=SQLDialect.SNOWFLAKE,
         pipeline=pipeline,
-        tables_columns={"beers_tiny": _BEERS_TABLE_COLUMNS},
+        tables_columns={"beers_tiny": BEERS_TABLE_COLUMNS},
         db_schema=None,
     )
     expected = pd.read_json(StringIO(json.dumps(pipeline_spec["expected"])), orient="table")
+    expected_rounded = expected.round(5)
+
     result: pd.DataFrame = pd.read_sql(text(query), engine)
-    assert_dataframes_equals(expected, result)
+    sanitized_result = _sanitize_datetimes(result)
+    result_rounded = sanitized_result.round(5)
+
+    assert_dataframes_equals(expected_rounded, result_rounded)