+++ server/CHANGELOG.md
@@ -2,11 +2,15 @@
 
 ## Unreleased
 
+### Fixed
+
+- PyPika: The Snowflake translator has been refreshed and its CI has been reactivated
+
 ## [0.57.0] - 2025-05-16
 
 ### Fixed
 
-- Mongo: `join` step adds also new columns to rows that do not match join conditions.   
+- Mongo: `join` step adds also new columns to rows that do not match join conditions.
 
 ### Changed